temporary file deletion

+++ sc4mpserver.py
@@ -1151,13 +1151,12 @@ class RegionsManager(th.Thread):
 
 					else:
 
-						# Clean up inbound temporary files and outputs
+						# Delete old temporary files
 						try:
 							path = os.path.join("_SC4MP", os.path.join("_Temp", "inbound"))
 							for directory in os.listdir(path):
-								if (directory in self.outputs.keys()):
+								if ((not directory in self.tasks) and (not directory in self.outputs.keys())):
 									shutil.rmtree(os.path.join(path, directory))
-									self.outputs.pop(directory)
 						except Exception as e:
 							pass
 						
@@ -1483,7 +1482,7 @@ class RequestHandler(th.Thread):
 					time.sleep(SC4MP_DELAY)
 
 				# Send the output to the client
-				c.send((sc4mp_regions_manager.outputs[save_id]).encode())
+				c.send((sc4mp_regions_manager.outputs.pop(save_id)).encode())
 
 			else:
 
@@ -1495,10 +1494,10 @@ class RequestHandler(th.Thread):
 			new_savegames = None
 
 		# Try to delete temporary files
-		try:
-			shutil.rmtree(path)
-		except:
-			pass
+		#try:
+		#	shutil.rmtree(path)
+		#except:
+		#	pass
 
 
 	def add_server(self, c):
